Fix: validate-build stages files in git prior to cleaning. That way only .gitignore files will be removed (i.e. built files)

+++ scripts/validate-build.ps1
@@ -1,5 +1,9 @@
 $ErrorActionPreference = 'Stop'
 
+Write-Host 'Staging untracked files...'
+git add -A
+if ($LASTEXITCODE -ne 0) { exit $LASTEXITCODE }
+
 Write-Host 'Cleaning src projects...'
 cd "$PSScriptRoot/../src"
 git clean -xdf
